Crate script_layout_interface

This module contains traits in script used generically in the rest of Servo. The traits are here instead of in script so that these modules won't have to depend on script.

Modules

message

rpc

wrapper_traits

Structs

DomParallelInfo
Information that we need stored in each DOM node.

HTMLCanvasData

HTMLMediaData

PendingImage
The data associated with an image that is not yet present in the image cache. Used by the script thread to hold on to DOM elements that need to be repainted when an image fetch is complete.

SVGSVGData

StyleAndGenericData

StyleData

TrustedNodeAddress
The address of a node known to be valid. These are sent from script to layout.

Enums

HTMLCanvasDataSource

LayoutElementType

LayoutNodeType

PendingImageState
Whether the pending image needs to be fetched or is waiting on an existing fetch.

Type Definitions

StyleAndOpaqueLayoutData