Struct script_layout_interface::message::Reflow

pub struct Reflow {
    pub goal: ReflowGoal,
    pub page_clip_rect: Rect<Au>,
}

Information needed for a reflow.

Fields
goal: ReflowGoal

The goal of reflow: either to render to the screen or to flush layout info for script.

page_clip_rect: Rect<Au>

A clipping rectangle for the page, an enlarged rectangle containing the viewport.