Struct script_layout_interface::PartialPersistentLayoutData

pub struct PartialPersistentLayoutData {
    pub style_data: NodeData,
    pub restyle_damage: RestyleDamage,
    pub parallel: DomParallelInfo,
}

Fields
style_data: NodeData

Data that the style system associates with a node. When the style system is being used standalone, this is all that hangs off the node. This must be first to permit the various transmutations between NodeData and PersistentLayoutData.

restyle_damage: RestyleDamage

Description of how to account for recent style changes.

parallel: DomParallelInfo

Information needed during parallel traversals.

Methods

impl PartialPersistentLayoutData

fn new() -> Self