Trait script_layout_interface::Layout

pub trait Layout {
    // Required methods
    fn process(&mut self, msg: Msg);
    fn handle_constellation_msg(&mut self, msg: LayoutControlMsg);
    fn handle_font_cache_msg(&mut self);
    fn rpc(&self) -> Box<dyn LayoutRPC>;
    fn waiting_for_web_fonts_to_load(&self) -> bool;
    fn current_epoch(&self) -> Epoch;
    fn load_web_fonts_from_stylesheet(&self, stylesheet: ServoArc<Stylesheet>);
    fn add_stylesheet(
        &mut self,
        stylesheet: ServoArc<Stylesheet>,
        before_stylsheet: Option<ServoArc<Stylesheet>>
    );
    fn remove_stylesheet(&mut self, stylesheet: ServoArc<Stylesheet>);
}

Required Methods

fn process(&mut self, msg: Msg)

Process a single message from script.

fn handle_constellation_msg(&mut self, msg: LayoutControlMsg)

Handle a single message from the Constellation.

fn handle_font_cache_msg(&mut self)

Handle a a single mesasge from the FontCacheThread.

fn rpc(&self) -> Box<dyn LayoutRPC>

Return the interface used for scipt queries. TODO: Make this part of the the Layout interface itself now that the layout thread has been removed.

fn waiting_for_web_fonts_to_load(&self) -> bool

Whether or not this layout is waiting for fonts from loaded stylesheets to finish loading.

fn current_epoch(&self) -> Epoch

The currently laid out Epoch that this Layout has finished.

fn load_web_fonts_from_stylesheet(&self, stylesheet: ServoArc<Stylesheet>)

Load all fonts from the given stylesheet, returning the number of fonts that need to be loaded.

fn add_stylesheet( &mut self, stylesheet: ServoArc<Stylesheet>, before_stylsheet: Option<ServoArc<Stylesheet>> )

Add a stylesheet to this Layout. This will add it to the Layout’s Stylist as well as loading all web fonts defined in the stylesheet. The second stylesheet is the insertion point (if it exists, the sheet needs to be inserted before it).

fn remove_stylesheet(&mut self, stylesheet: ServoArc<Stylesheet>)

Removes a stylesheet from the Layout.