Trait script_layout_interface::wrapper_traits::LayoutNode

pub trait LayoutNode: Debug + GetLayoutData + TNode {
    type ConcreteThreadSafeLayoutNode: ThreadSafeLayoutNode;
    fn to_threadsafe(&self) -> Self::ConcreteThreadSafeLayoutNode;
    fn type_id(&self) -> LayoutNodeType;
    unsafe fn init_style_and_layout_data(&self, data: OpaqueStyleAndLayoutData);
    unsafe fn take_style_and_layout_data(&self) -> OpaqueStyleAndLayoutData;
    fn first_child(&self) -> Option<Self>;
    fn last_child(&self) -> Option<Self>;
    fn prev_sibling(&self) -> Option<Self>;
    fn next_sibling(&self) -> Option<Self>;

    fn rev_children(self) -> LayoutIterator<ReverseChildrenIterator<Self>> { ... }
    fn traverse_preorder(self) -> TreeIterator<Self> { ... }
}

A wrapper so that layout can access only the methods that it should have access to. Layout must only ever see these and must never see instances of LayoutJS.

Associated Types

type ConcreteThreadSafeLayoutNode: ThreadSafeLayoutNode

Required Methods

fn to_threadsafe(&self) -> Self::ConcreteThreadSafeLayoutNode

fn type_id(&self) -> LayoutNodeType

Returns the type ID of this node.

unsafe fn init_style_and_layout_data(&self, data: OpaqueStyleAndLayoutData)

unsafe fn take_style_and_layout_data(&self) -> OpaqueStyleAndLayoutData

fn first_child(&self) -> Option<Self>

fn last_child(&self) -> Option<Self>

fn prev_sibling(&self) -> Option<Self>

fn next_sibling(&self) -> Option<Self>

Provided Methods

fn rev_children(self) -> LayoutIterator<ReverseChildrenIterator<Self>>

fn traverse_preorder(self) -> TreeIterator<Self>