Crate script_layout_interface

This module contains traits in script used generically in the rest of Servo. The traits are here instead of in script so that these modules won't have to depend on script.

Modules

message

reporter

restyle_damage

rpc

wrapper_traits

Structs

HTMLCanvasData

OpaqueStyleAndLayoutData

PartialStyleAndLayoutData

TrustedNodeAddress
The address of a node known to be valid. These are sent from script to layout.

Enums

LayoutElementType

LayoutNodeType

Functions

is_image_data