Crate script_layout_interface

This module contains traits in script used generically in the rest of Servo. The traits are here instead of in script so that these modules won’t have to depend on script.

Modules

wrapper_traits

Structs

DomParallelInfo

Information that we need stored in each DOM node.

HTMLCanvasData

HTMLMediaData

LayoutConfig

OffsetParentResponse

PendingImage

The data associated with an image that is not yet present in the image cache. Used by the script thread to hold on to DOM elements that need to be repainted when an image fetch is complete.

PendingRestyle

A pending restyle.

Reflow

Information needed for a reflow.

ReflowComplete

Information derived from a layout pass that needs to be returned to the script thread.

SVGSVGData

ScriptReflow

Information needed for a script-initiated reflow.

StyleData

TrustedNodeAddress

The address of a node known to be valid. These are sent from script to layout.

Enums

FragmentType

The type of fragment that a scroll root is created for.

HTMLCanvasDataSource

LayoutElementType

LayoutNodeType

NodesFromPointQueryType

PendingImageState

Whether the pending image needs to be fetched or is waiting on an existing fetch.

QueryMsg

ReflowGoal

Any query to perform with this reflow.

Constants

SPECIAL_SCROLL_ROOT_ID_MASK 🔒

If none of the bits outside this mask are set, the scroll root is a special scroll root. Note that we assume that the top 16 bits of the address space are unused on the platform.

Statics

NEXT_SPECIAL_SCROLL_ROOT_ID 🔒

The next ID that will be used for a special scroll root id.

Traits

Layout

LayoutFactory

ScriptThreadFactory

This trait is part of script_layout_interface because it depends on both script_traits and also LayoutFactory from this crate. If it was in script_traits there would be a circular dependency.

Functions

combine_id_with_fragment_type

next_special_id 🔒

Returns a new scroll root ID for a scroll root.

node_id_from_scroll_id

Type Aliases

GenericLayoutData