Trait script_layout_interface::wrapper_traits::LayoutNode

pub trait LayoutNode<'dom>: Debug + GetStyleAndOpaqueLayoutData<'dom> + TNode {
    type ConcreteThreadSafeLayoutNode: ThreadSafeLayoutNode<'dom>;
    fn to_threadsafe(&self) -> Self::ConcreteThreadSafeLayoutNode;

    fn type_id(&self) -> LayoutNodeType;

    unsafe fn initialize_data(&self);

    unsafe fn init_style_and_opaque_layout_data(
        &self, 
        data: Box<StyleAndOpaqueLayoutData>
    );

    unsafe fn take_style_and_opaque_layout_data(
        &self
    ) -> Box<StyleAndOpaqueLayoutData>;

    fn is_connected(&self) -> bool;

    fn rev_children(self) -> LayoutIterator<ReverseChildrenIterator<Self>> { ... }

    fn traverse_preorder(self) -> TreeIterator<Self>ⓘ

Important traits for TreeIterator<ConcreteNode>
impl<'dom, ConcreteNode> Iterator for TreeIterator<ConcreteNode> where
    ConcreteNode: LayoutNode<'dom>,     type Item = ConcreteNode;
 { ... }
}

A wrapper so that layout can access only the methods that it should have access to. Layout must only ever see these and must never see instances of LayoutDom.

Associated Types

type ConcreteThreadSafeLayoutNode: ThreadSafeLayoutNode<'dom>

Loading content...
Required methods

fn to_threadsafe(&self) -> Self::ConcreteThreadSafeLayoutNode

fn type_id(&self) -> LayoutNodeType

Returns the type ID of this node.

unsafe fn initialize_data(&self)

unsafe fn init_style_and_opaque_layout_data(
&self,
data: Box<StyleAndOpaqueLayoutData>
)

unsafe fn take_style_and_opaque_layout_data(
&self
) -> Box<StyleAndOpaqueLayoutData>

fn is_connected(&self) -> bool

Returns whether the node is connected.

Loading content...
Provided methods

fn rev_children(self) -> LayoutIterator<ReverseChildrenIterator<Self>>

fn traverse_preorder(self) -> TreeIterator<Self>ⓘ

Important traits for TreeIterator<ConcreteNode>
impl<'dom, ConcreteNode> Iterator for TreeIterator<ConcreteNode> where
ConcreteNode: LayoutNode<'dom>, type Item = ConcreteNode;

Loading content...