Trait script_layout_interface::wrapper_traits::ThreadSafeLayoutElement

pub trait ThreadSafeLayoutElement: Clone + Copy + Sized + Debug + Element<Impl=ServoSelectorImpl> + GetLayoutData + PresentationalHintsSynthetizer {
    type ConcreteThreadSafeLayoutNode: ThreadSafeLayoutNode<ConcreteThreadSafeLayoutElement=Self>;
    fn as_node(&self) -> Self::ConcreteThreadSafeLayoutNode;
    fn with_pseudo(&self, pseudo: PseudoElementType<Option<T>>) -> Self;
    fn type_id(&self) -> Option<LayoutNodeType>;
    fn get_attr(&self, namespace: &Namespace, name: &LocalName) -> Option<&str>;
    fn get_style_data(&self) -> Option<&AtomicRefCell<ElementData>>;
    fn get_pseudo_element_type(&self) -> PseudoElementType<Option<T>>;

    fn get_before_pseudo(&self) -> Option<Self> { ... }
    fn get_after_pseudo(&self) -> Option<Self> { ... }
    fn get_details_summary_pseudo(&self) -> Option<Self> { ... }
    fn get_details_content_pseudo(&self) -> Option<Self> { ... }
    fn style(&self, context: &SharedStyleContext) -> Arc<ServoComputedValues> { ... }
    fn selected_style(&self) -> Arc<ServoComputedValues> { ... }
    fn resolved_style(&self) -> Arc<ServoComputedValues> { ... }
}

Associated Types

type ConcreteThreadSafeLayoutNode: ThreadSafeLayoutNode<ConcreteThreadSafeLayoutElement=Self>

Required Methods

fn as_node(&self) -> Self::ConcreteThreadSafeLayoutNode

fn with_pseudo(&self, pseudo: PseudoElementType<Option<T>>) -> Self

Creates a new ThreadSafeLayoutElement for the same LayoutElement with a different pseudo-element type.

fn type_id(&self) -> Option<LayoutNodeType>

Returns the type ID of this node. Returns None if this is a pseudo-element; otherwise, returns Some.

fn get_attr(&self, namespace: &Namespace, name: &LocalName) -> Option<&str>

fn get_style_data(&self) -> Option<&AtomicRefCell<ElementData>>

fn get_pseudo_element_type(&self) -> PseudoElementType<Option<T>>

Provided Methods

fn get_before_pseudo(&self) -> Option<Self>

fn get_after_pseudo(&self) -> Option<Self>

fn get_details_summary_pseudo(&self) -> Option<Self>

fn get_details_content_pseudo(&self) -> Option<Self>

fn style(&self, context: &SharedStyleContext) -> Arc<ServoComputedValues>

Returns the style results for the given node. If CSS selector matching has not yet been performed, fails.

Unlike the version on TNode, this handles pseudo-elements.

fn selected_style(&self) -> Arc<ServoComputedValues>

fn resolved_style(&self) -> Arc<ServoComputedValues>

Returns the already resolved style of the node.

This differs from style(ctx) in that if the pseudo-element has not yet been computed it would panic.

This should be used just for querying layout, or when we know the element style is precomputed, not from general layout itself.