Trait script_layout_interface::wrapper_traits::ThreadSafeLayoutNode

pub trait ThreadSafeLayoutNode: Clone + Copy + GetLayoutData + NodeInfo + PartialEq + Sized {
    type ConcreteThreadSafeLayoutElement: ThreadSafeLayoutElement<ConcreteThreadSafeLayoutNode=Self> + Element<Impl=ServoSelectorImpl>;
    type ChildrenIterator: Iterator<Item=Self> + Sized;
    fn opaque(&self) -> OpaqueNode;
    fn type_id(&self) -> Option<LayoutNodeType>;
    fn type_id_without_excluding_pseudo_elements(&self) -> LayoutNodeType;
    fn style_for_text_node(&self) -> Arc<ServoComputedValues>;
    fn debug_id(self) -> usize;
    fn children(&self) -> LayoutIterator<Self::ChildrenIterator>;
    fn as_element(&self) -> Option<Self::ConcreteThreadSafeLayoutElement>;
    fn get_style_and_layout_data(&self) -> Option<OpaqueStyleAndLayoutData>;
    fn is_ignorable_whitespace(&self, context: &SharedStyleContext) -> bool;
    fn restyle_damage(self) -> RestyleDamage;
    fn clear_restyle_damage(self);
    fn can_be_fragmented(&self) -> bool;
    fn node_text_content(&self) -> String;
    fn selection(&self) -> Option<Range<ByteIndex>>;
    fn image_url(&self) -> Option<Url>;
    fn canvas_data(&self) -> Option<HTMLCanvasData>;
    fn svg_data(&self) -> Option<SVGSVGData>;
    fn iframe_pipeline_id(&self) -> PipelineId;
    fn get_colspan(&self) -> u32;

    fn is_element_or_elements_pseudo(&self) -> bool { ... }
    fn get_before_pseudo(&self) -> Option<Self> { ... }
    fn get_after_pseudo(&self) -> Option<Self> { ... }
    fn get_details_summary_pseudo(&self) -> Option<Self> { ... }
    fn get_details_content_pseudo(&self) -> Option<Self> { ... }
    fn get_pseudo_element_type(&self) -> PseudoElementType<Option<T>> { ... }
    fn style(&self, context: &SharedStyleContext) -> Arc<ServoComputedValues> { ... }
    fn selected_style(&self) -> Arc<ServoComputedValues> { ... }
    fn is_content(&self) -> bool { ... }
}

A thread-safe version of LayoutNode, used during flow construction. This type of layout node does not allow any parents or siblings of nodes to be accessed, to avoid races.

Associated Types

type ConcreteThreadSafeLayoutElement: ThreadSafeLayoutElement<ConcreteThreadSafeLayoutNode=Self> + Element<Impl=ServoSelectorImpl>

type ChildrenIterator: Iterator<Item=Self> + Sized

Required Methods

fn opaque(&self) -> OpaqueNode

Converts self into an OpaqueNode.

fn type_id(&self) -> Option<LayoutNodeType>

Returns the type ID of this node. Returns None if this is a pseudo-element; otherwise, returns Some.

fn type_id_without_excluding_pseudo_elements(&self) -> LayoutNodeType

Returns the type ID of this node, without discarding pseudo-elements as type_id does.

fn style_for_text_node(&self) -> Arc<ServoComputedValues>

Returns the style for a text node. This is computed on the fly from the parent style to avoid traversing text nodes in the style system.

Note that this does require accessing the parent, which this interface technically forbids. But accessing the parent is only unsafe insofar as it can be used to reach siblings and cousins. A simple immutable borrow of the parent data is fine, since the bottom-up traversal will not process the parent until all the children have been processed.

fn debug_id(self) -> usize

fn children(&self) -> LayoutIterator<Self::ChildrenIterator>

Returns an iterator over this node's children.

fn as_element(&self) -> Option<Self::ConcreteThreadSafeLayoutElement>

Returns a ThreadSafeLayoutElement if this is an element, None otherwise.

fn get_style_and_layout_data(&self) -> Option<OpaqueStyleAndLayoutData>

fn is_ignorable_whitespace(&self, context: &SharedStyleContext) -> bool

fn restyle_damage(self) -> RestyleDamage

fn clear_restyle_damage(self)

fn can_be_fragmented(&self) -> bool

fn node_text_content(&self) -> String

fn selection(&self) -> Option<Range<ByteIndex>>

If the insertion point is within this node, returns it. Otherwise, returns None.

fn image_url(&self) -> Option<Url>

If this is an image element, returns its URL. If this is not an image element, fails.

FIXME(pcwalton): Don't copy URLs.

fn canvas_data(&self) -> Option<HTMLCanvasData>

fn svg_data(&self) -> Option<SVGSVGData>

fn iframe_pipeline_id(&self) -> PipelineId

If this node is an iframe element, returns its pipeline ID. If this node is not an iframe element, fails.

fn get_colspan(&self) -> u32

Provided Methods

fn is_element_or_elements_pseudo(&self) -> bool

fn get_before_pseudo(&self) -> Option<Self>

fn get_after_pseudo(&self) -> Option<Self>

fn get_details_summary_pseudo(&self) -> Option<Self>

fn get_details_content_pseudo(&self) -> Option<Self>

fn get_pseudo_element_type(&self) -> PseudoElementType<Option<T>>

fn style(&self, context: &SharedStyleContext) -> Arc<ServoComputedValues>

fn selected_style(&self) -> Arc<ServoComputedValues>

fn is_content(&self) -> bool

Returns true if this node contributes content. This is used in the implementation of empty_cells per CSS 2.1  17.6.1.1.